Completed all formatting
Added the new package needed (will probably rename it)
Moved the ProcessActions class into the new package.
Renamed ProcessConstructor to FileConstructor .
