+++ HP like game/src/com/crazykid080/main/actions/ProcessActions.java
@@ -1,0 +1,27 @@
+package com.crazykid080.main.actions;
+
+import com.crazykid080.main.constructors.FileConstructor;
+import com.crazykid080.main.userControl.Server;
+
+public class ProcessActions{
+
+	public void onProcessKill(FileConstructor obj, boolean log){
+
+		obj.setRunning(false);
+
+		if(log == true){
+			Server s = obj.getServer();
+			s.writeLog(" killed process ");
+		}
+
+	}
+
+	public void onProcessStart(FileConstructor obj, boolean log){
+
+	}
+
+	public void onInstallStart(/*NEED TO CREATE THE FILES*/ int mil){
+		//new timer(time, file);
+	}
+
+}
